requirements.txt
Flask>=3.0
numpy>=1.26
opencv-python>=4.10
Pillow>=10.0
scikit-learn>=1.5
torch>=2.4
torchvision>=0.19
ultralytics>=8.3
transformers>=4.45
pytesseract>=0.3.10
qiskit>=2.0
qiskit-machine-learning>=0.8
joblib>=1.4
python-pptx>=1.0
